requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
openai
jsonschema
python-multipart